refactor(hudi-common): move isTimestampMillisField method to HoodieSchemaUtils class
